+++ testing/e2e-test-utilities/src/main/kotlin/net/corda/e2etest/utilities/MGMUtils.kt
@@ -294,23 +294,6 @@ fun ClusterInfo.declineRegistration(
     }
 }
 
-internal enum class TlsType(
-    val groupPolicyMame: String,
-    val configName: String,
-) {
-    ONE_WAY("OneWay", "ONE_WAY"),
-    MUTUAL("Mutual", "MUTUAL");
-
-    companion object {
-        val type by lazy {
-            if (System.getenv("CORDA_E2E_TEST_TLS_TYPE") != null) {
-                MUTUAL
-            } else {
-                ONE_WAY
-            }
-        }
-    }
-}
 /**
  * Data class for customising the group policy file during MGM registration.
  */
